web - conta modal

+++ src/MoneyLoris.Application/Business/Contas/Dtos/ContaCriacaoDto.cs
@@ -9,10 +9,10 @@ public class ContaCriacaoDto
     public string Cor { get; set; } = default!;
     public int? Ordem { get; set; }
 
-    public double SaldoInicial { get; set; }
+    public decimal SaldoInicial { get; set; }
     public DateTime DataLancamentoInicial { get; set; }
 
-    public double Limite { get; set; }
+    public decimal Limite { get; set; }
     public byte DiaFechamento { get; set; }
     public byte DiaVencimento { get; set; }
 }
